{"schema":"decodex/commit/1","summary":"Split MCP lane control result modules","authority":"XY-1115"}
